requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastapi-utils==0.8.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
m3u8==6.0.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mega.py==1.0.8
multidict==6.1.0
numpy==2.1.3
orjson==3.10.10
passlib==1.7.4
pathlib==1.0.1
patreon==0.5.0
propcache==0.2.0
psutil==5.9.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.3
rsa==4.9
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
tenacity==5.1.5
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
async-timeout>=4.0,<5.0
python-dateutil
edge-tts==7.2.1
pypdf>=3.0.0
pdfplumber>=0.9.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
ebooklib>=0.18
python-docx>=0.8.11
msgpack==1.1.0

zstandard>=0.22.0
